+++ doc/CHANGELOG.md
@@ -28,6 +28,8 @@ For release instructions, see [cynic-net/pypi-release] on GitHub.
 - Changed: `MemImage` now automatically recalculates `startaddr` and
   `endaddr` when a record is added, and these can be overridden afterwards
   by the user.
+- Added: `p2b` program to generate generate Apple II "B" type and and KC85
+  .CO program files from ASL .p files.
 
 ### 0.0.7 (2024-09-22)
 - Fixed: `RomImage.patches()` now works if more than one patchspec matches.